python/query/requirements.txt
pandas==2.0.2
pandasql==0.7.3
grpcio==1.56.0
google==3.0.0
grpcio-tools==1.56.0
retrying==1.3.4